Towncrier news fragments for next releease.
(Forgot to make them during feature development.)

(Also, credit to Anthropic claude-3-5-sonnet for documentation and tests
on the 'mutables' and 'one' features. And, credit to this model and to
OpenAI o1 for helping me remove some type checker ignore comments on the
'ProtocolClass' metaclass. Forgot to mention on previous commits.)